Audio/Video detection
Use the avalable mime-type list to set if an mime-type contains a video
stream or not. libmpv can't detect this without going crazy if an video
stream is available but no surface to render into it (e.g. on some audiofiles
with cover image). For our luck, the most common container formats are
clearly defined as audio-only or not. The backend will now disable the
video output if the mime-type doesn't support it.