Fix possible crash when a file is done playing
There's a race condition in which MediaObject::event_cb can be called
(as a result of an event caught by mpv_set_wakeup_callback) after the
object pointed to by the opaque pointer has already been destroyed.

This causes phonon-mpv to crash when a file has finished playing.

This fix isn't very nice (essentially it just keeps a list of
recently destroyed MediaObjects and avoids calling invokeMethod on
them), but it works.